Split visuals of HitCircleSelectionMask into HitCircleMask
